Add react-cookie package to save the logged in user in a cookie and set up the admin and regular user's menu

+++ src/components/TopBar.jsx
@@ -1,5 +1,5 @@
 import React, { useContext, useState } from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useHistory } from 'react-router-dom';
 import { Navbar } from 'react-bulma-components';
 import { FiShoppingCart } from 'react-icons/fi';
 import axios from 'axios';
@@ -12,22 +12,29 @@ const TopBar = () => {
   const userCtx = useContext(UserContext);
   const cartCtx = useContext(CartContext);
   const numItems = cartCtx.cartCount;
+  const history = useHistory();
 
   const handleLogout = () => {
-    axios({
-      method: 'delete',
-      url: 'http://localhost:3000/auth/logout',
-      withCredentials: true,
-    }).then((res) => {
-      if (res.status === 200) {
-        userCtx.setUser({});
-        console.log(res.data);
-      }
-    })
-      .catch((err) => console.log(err));
+    if (userCtx.cookies.user) {
+      axios({
+        method: 'delete',
+        url: 'http://localhost:3000/auth/logout',
+        withCredentials: true,
+      })
+        .then((res) => {
+          if (res.status === 200) {
+            userCtx.removeCookie('user');
+            history.push('/auth');
+            console.log(res.data);
+          }
+        })
+        .catch((err) => console.log(err));
+    } else {
+      return console.log('login first');
+    }
   };
 
-
+  // console.log(userCtx.cookies.user)
   return (
 
     <Navbar
@@ -45,16 +52,16 @@ const TopBar = () => {
 
       <Navbar.Menu>
         {
-          Object.keys(userCtx.user).length > 0
+          (userCtx.cookies.user)
             ? (
               <Navbar.Container>
-                <div>
-                  <h5>
-                    Logged in as:
-                    {userCtx.user.email}
+                <div className="top-bar__user-div">
+                  <h5 className="top-bar__user-div--text">
+                    Logged in as:&nbsp;
+                    <span className="is-inline-block is-bold has-capital-fletter">{userCtx.cookies.user.username}</span>
                   </h5>
                   {
-                    userCtx.user.admin === true ? <h5>Admin</h5> : null
+                    userCtx.cookies.user.admin === true ? <h5 className="top-bar__user-div--text is-bold">Admin</h5> : null
                   }
                 </div>
               </Navbar.Container>
@@ -79,15 +86,20 @@ const TopBar = () => {
               </span>
             </Link>
           </div>
+          {
+            (userCtx.cookies.user && userCtx.cookies.user.admin)
+              ? (
+                <div className="top-bar__link-div">
+                  <Link to="/add" className="top-bar__link is-bold">
+                    Add
+                  </Link>
+                </div>
+              )
+              : null
+          }
 
           <div className="top-bar__link-div">
-            <Link to="/add" className="top-bar__link is-bold">
-              Add
-            </Link>
-          </div>
-
-          <div className="top-bar__link-div">
-            { Object.keys(userCtx.user).length < 1
+            { (!userCtx.cookies.user)
               ? (
                 <Link to="/auth" className="top-bar__link is-bold">
                   Login
